requirements.txt
pytz>=2016.6.1
luigi>=2.2.0
pyquery>=1.2.9
requests>=2.9.1
pymongo>=3.2
python-dateutil>=2.4.2
pandas>=0.17.1
argparse>=1.1.0